add the model, refactor the Project
