+++ Artikel/src/main/java/de/ProPra/Articles/VerleihApplication.java
@@ -1,4 +1,4 @@
-package de.ProPra.Verleih;
+package de.ProPra.Articles;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
